fix: move up enviroments to src
